requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
beautifulsoup4
apscheduler
openai
python-dotenv
pydantic
sse-starlette
google-genai
feedparser
